fix the dto

+++ Models/Rating/Rating.cs
@@ -3,7 +3,7 @@ namespace CreationHub.Models;
 public class Rating
 {
     public long Id { get; set; }
-    public int NicePartUsageId { get; set; }
+    public long NicePartUsageId { get; set; }
     public int? Creativity { get; set; }
 
     public int? Uniqueness { get; set; }
